Fix namespacing of concerns
