+++ app/controllers/concerns/displays_navigation.rb
@@ -1,0 +1,7 @@
+module DisplaysNavigation
+  extend ActiveSupport::Concern
+
+  def disable_navigation
+    @disable_navigation = true
+  end
+end
