--- app/controllers/concerns/navigation.rb
@@ -1,9 +1,0 @@
-module Concerns
-  module Navigation
-    extend ActiveSupport::Concern
-
-    def disable_navigation
-      @disable_navigation = true
-    end
-  end
-end
